refactor(api): restore route.ts and prompt.ts from main
